fix test on spark2
